[controller] Retry topic check during store migration (#2337)
During store migration we check for largest used version in src colo and topic existence to determine whether to skip migrating that version or not. But if due to kafka metadata delay the topic existence check returns false, we could miss migrating a valid version. This PR adds retries to the topic check and added logs to help debug.