[server] Remove incorrect error log/metric of RMD timestamp validation (#2351)
This log is very noisy and is incorrect as today we don't have complete implementation of field level RMD timestamp validation post operation, and it is default to 0. When ever a value level TS transitioned into field level TS (PUT->UPDATE)
then it will log this line with metric recording.

Before the actual measurement is correctly implemented remove this one to remove the noisy logs.